moving parallel attributes
