+++ src/SeleniumSharpTemplate/Utilities/Attributes/IParallelAttributes.cs
@@ -1,7 +1,7 @@
 ﻿[assembly: LevelOfParallelism(4)]
 [assembly: Parallelizable(ParallelScope.Fixtures)]
 
-namespace SeleniumSharpTemplate
+namespace SeleniumSharpTemplate.Utilities.Attributes
 {
     internal interface IParallelAttributes;
 }
