+++ src/SeleniumSharpTemplate/Utilities/Logging/Listeners/DriverEventListeners.cs
@@ -1,4 +1,5 @@
 ﻿using OpenQA.Selenium.Support.Events;
+using SeleniumSharpTemplate.Utilities.DateTime;
 
 namespace SeleniumSharpTemplate.Utilities.Logging.Events
 {
@@ -7,35 +8,37 @@ namespace SeleniumSharpTemplate.Utilities.Logging.Events
     {
         private const string PASSWORD_MASK = "********";
 
+        private static string EVENT => DateTimeFormatter.NOW_YYYY_MM_DD_HH_MM_SS_UTC + " [EVENT]";
+
         private static readonly Dictionary<IWebElement, string> WebElementNames = [];
 
         private static readonly List<string> ListOfAttributes = ["value", "name", "id", "placeholder", "class"];
 
         public static void ExceptionThrown(object? sender, WebDriverExceptionEventArgs eventArgs) =>
-            Console.WriteLine($"Exception: {eventArgs.ThrownException.Message}");
+            Console.WriteLine($"{EVENT} Exception: {eventArgs.ThrownException.Message}");
 
         public static void ElementClicking(object? sender, WebElementEventArgs eventArgs) =>
             WebElementNames[eventArgs.Element] = ElementProperty(eventArgs.Element);
 
         public static void ElementClicked(object? sender, WebElementEventArgs eventArgs) =>
-            Console.WriteLine($"Element '{WebElementNames[eventArgs.Element]}' is clicked");
+            Console.WriteLine($"{EVENT} Element '{WebElementNames[eventArgs.Element]}' is clicked");
 
         public static void ElementValueChanged(object? sender, WebElementValueEventArgs eventArgs)
         {
             WebElementNames[eventArgs.Element] = eventArgs.Element.GetAttribute("name");
 
             if (eventArgs.Value != null)
-                Console.WriteLine($"Element '{eventArgs.Element.GetAttribute("name")}' value changed from '{WebElementNames[eventArgs.Element]}' to: '{ElementProperty(eventArgs.Element)}'");
+                Console.WriteLine($"{EVENT} Element '{eventArgs.Element.GetAttribute("name")}' value changed from '{WebElementNames[eventArgs.Element]}' to: '{ElementProperty(eventArgs.Element)}'");
         }
 
         public static void FindElementCompleted(object? sender, FindElementEventArgs eventArgs) =>
-            Console.WriteLine($"Found element: '{eventArgs.FindMethod}'");
+            Console.WriteLine($"{EVENT} Found element: '{eventArgs.FindMethod}'");
 
         public static void Navigated(object? sender, WebDriverNavigationEventArgs eventArgs) =>
-            Console.WriteLine($"Navigated to: '{eventArgs.Driver.Title}'");
+            Console.WriteLine($"{EVENT} Navigated to: '{eventArgs.Driver.Title}'");
 
         public static void ScriptExecuted(object? sender, WebDriverScriptEventArgs eventArgs) =>
-            Console.WriteLine($"Script '{eventArgs.Script.Split('.')[1].Split('(')[0]}' executed: '{eventArgs.Script}'");
+            Console.WriteLine($"{EVENT} Script '{eventArgs.Script.Split('.')[1].Split('(')[0]}' executed: '{eventArgs.Script}'");
 
         private static string ElementProperty(IWebElement element)
         {
